integration: tests: break into multiple modules
