+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "contracts-utils",
     "integration",
 ]
+resolver = "2"
 
 [workspace.dependencies]
 ark-bn254 = "0.4.0"
